Move payload handling code out of controller
lays the framework for cleaning this up. What we should do
is just have this create a single handling job, and then that
job splits and delegates up the work into multiple jobs. This
keeps the incoming webhook call from github exposed to the smallest
possible surface area for errors